+++ frontend/src/components/index.ts
@@ -8,6 +8,8 @@ export * from 'components/@shared/InfiniteScroll';
 export * from 'components/@shared/BottomSheet';
 export * from 'components/@shared/Dropdown';
 export * from 'components/@shared/Input';
+export * from 'components/@shared/LoadingButton';
+export * from 'components/@shared/LoadingSpinner';
 export * from 'components/LinkText';
 export * from 'components/Logo';
 export * from 'components/ValidationMessage';
@@ -23,12 +25,10 @@ export * from 'components/ChallengeIcon';
 export * from 'components/Profile';
 export * from 'components/EmptyContent';
 export * from 'components/SnackBar';
-export * from 'components/LoadingSpinner';
 export * from 'components/PrivateOutlet';
 export * from 'components/LandingNavigation';
 export * from 'components/ToggleButton';
 export * from 'components/Title';
-export * from 'components/LoadingButton';
 export * from 'components/Challenger';
 export * from 'components/ChallengerList';
 export * from 'components/ChallengeDetail';
